# src/main/resources/swagger/loja.yaml
...
swagger: "2.0"
info:
  description: "Sistema de Loja"
  version: "1.0.0"
  title: "Loja"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8082"
basePath: "/v1"
...
schemes:
  - "https"
  - "http"
paths:
  /categoria:
    put:
      tags:
        - "Sistema Loja"
      summary: "atualizar categoria"
      description: "Atualizar Categoria da loja"
      operationId: "atualizarCategoria"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Categoria object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        200:
          description: "successful operation"
        201:
          description: "Created"
        400:
          description: "Bad Request"
        404:
          description: "Not Found"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ApiResponse"

    post:
      tags:
        - "Sistema Loja"
      summary: "Add a new categoria to the store"
      description: ""
      operationId: "addCategoria"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Categoria object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        405:
          description: "Invalid input"
        201:
          description: "Created"
        400:
          description: "Bad Request"
        404:
          description: "Not Found"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ApiResponse"
...
  /user:
    post:
      tags:
        - "user"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
        - "application/xml"
      parameters:
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "https://petstore.swagger.io/oauth/authorize"
    flow: "implicit"
    scopes:
      read:pets: "read your pets"
      write:pets: "modify pets in your account"
definitions:
  ApiResponse:
    type: "object"
    properties:
      title:
        type: "string"
      message:
        type: "string"
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
    xml:
      name: "User"
...